vu: 64bit cycle fixes
[SAVEVERSION+]

+++ pcsx2/VU0microInterp.cpp
@@ -64,7 +64,7 @@ static void _vu0Exec(VURegs* VU)
 	VU->code = ptr[1];
 	VU0regs_UPPER_OPCODE[VU->code & 0x3f](&uregs);
 
-	u32 cyclesBeforeOp = VU0.cycle - 1;
+	u64 cyclesBeforeOp = VU0.cycle - 1;
 
 	_vuTestUpperStalls(VU, &uregs);
 
@@ -252,7 +252,7 @@ void InterpVU0::Execute(u32 cycles)
 
 	VU0.VI[REG_TPC].UL <<= 3;
 	VU0.flags &= ~VUFLAG_MFLAGSET;
-	u32 startcycles = VU0.cycle;
+	u64 startcycles = VU0.cycle;
 	while ((VU0.cycle - startcycles) < cycles)
 	{
 		if (!(VU0.VI[REG_VPU_STAT].UL & 0x1))
@@ -274,7 +274,7 @@ void InterpVU0::Execute(u32 cycles)
 
 	if (EmuConfig.Speedhacks.EECycleRate != 0 && (!EmuConfig.Gamefixes.VUSyncHack || EmuConfig.Speedhacks.EECycleRate < 0))
 	{
-		u32 cycle_change = VU0.cycle - startcycles;
+		u64 cycle_change = VU0.cycle - startcycles;
 		VU0.cycle -= cycle_change;
 		switch (std::min(static_cast<int>(EmuConfig.Speedhacks.EECycleRate), static_cast<int>(cycle_change)))
 		{
