+++ pcsx2/VU1microInterp.cpp
@@ -261,7 +261,7 @@ void InterpVU1::Execute(u32 cycles)
 	const FPControlRegisterBackup fpcr_backup(EmuConfig.Cpu.VU1FPCR);
 
 	VU1.VI[REG_TPC].UL <<= 3;
-	u32 startcycles = VU1.cycle;
+	u64 startcycles = VU1.cycle;
 
 	while ((VU1.cycle - startcycles) < cycles)
 	{
